Error solucionado

+++ services/server/product.ts
@@ -4,15 +4,30 @@ import { supabase } from "@/config/supabase";
 
 //funcion para eliminar productos de la tabla (product)
 export async function deleteProduct(id : string) {
-    const { error } = await supabase
+    //Elimina el registro relacionado con product_supplier
+    const { error: errorSupplier } = await supabase
+        .from('product_supplier')
+        .delete()
+        .eq('id_product', id);
+        
+    //Manejo de errores
+    if (errorSupplier) {
+        console.log('Error:', errorSupplier.message);
+        return { error: errorSupplier }
+    } 
+    //Eliminar el producto
+    const { error: errorProduct } = await supabase
         .from('product')
         .delete()
         .eq('id', id);
-        
+    
     //Manejo de errores
-    if (error) {
-        console.log('Error:', error.message);
-    } else {
-        console.log('Producto eliminado');
+    if (errorProduct) {
+        console.log('Error al eliminar el producto:', errorProduct.message);
+        return { error: errorProduct }
+    }else{
+        console.log('Producto eliminado exitosamente.');
+        return { error: null }
     }
+    
 }
